@@ -1,4 +1,4 @@
-// swift-tools-version:5.5
+// swift-tools-version:5.6
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -7,13 +7,10 @@ let package = Package(
     name: "Get",
     platforms: [.iOS(.v13), .macCatalyst(.v13), .macOS(.v10_15), .watchOS(.v6), .tvOS(.v13)],
     products: [
-        .library(name: "Get", targets: ["Get"]),
-    ],
-    dependencies: [
-        .package(url: "https://github.com/WeTransfer/Mocker.git", from: "2.5.6")
+        .library(name: "Get", targets: ["Get"])
     ],
     targets: [
         .target(name: "Get"),
-        .testTarget(name: "GetTests", dependencies: ["Get", "Mocker"], resources: [.process("Resources")]),
+        .testTarget(name: "GetTests", dependencies: ["Get"], resources: [.process("Resources")])
     ]
 )
