@@ -8,27 +8,37 @@ import Foundation
 import FoundationNetworking
 #endif
 
-public struct Request<Response> {
+/// An HTTP network request.
+public struct Request<Response>: @unchecked Sendable {
+    /// HTTP method, e.g. "GET".
     public var method: String
+    /// Resource path.
     public var path: String
+    /// Request query items.
     public var query: [(String, String?)]?
-    var body: AnyEncodable?
+    /// Request headers to be added to the request.
     public var headers: [String: String]?
+    /// ID provided by the user. Not used by the API client.
     public var id: String?
 
+    let body: AnyEncodable?
+
+    /// Initialiazes the request with the given parameters.
     public init(method: String, path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) {
         self.method = method
         self.path = path
         self.query = query
         self.headers = headers
+        self.body = nil
     }
 
+    /// Initialiazes the request with the given parameters and the request body.
     public init<U: Encodable>(method: String, path: String, query: [(String, String?)]? = nil, body: U?, headers: [String: String]? = nil) {
         self.method = method
         self.path = path
         self.query = query
-        self.body = body.map(AnyEncodable.init)
         self.headers = headers
+        self.body = body.map(AnyEncodable.init)
     }
 
     public static func get(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
@@ -80,31 +90,14 @@ public struct Request<Response> {
     }
 }
 
-/// A response with a value and associated metadata.
-public struct Response<T> {
-    public var value: T
-    /// Original response data.
-    public var data: Data
-    /// Original request.
-    public var request: URLRequest
-    public var response: URLResponse
-    public var statusCode: Int? { (response as? HTTPURLResponse)?.statusCode }
-    public var metrics: URLSessionTaskMetrics?
+struct AnyEncodable: Encodable {
+    private let value: Encodable
 
-    public init(value: T, data: Data, request: URLRequest, response: URLResponse, metrics: URLSessionTaskMetrics? = nil) {
+    init(_ value: Encodable) {
         self.value = value
-        self.data = data
-        self.request = request
-        self.response = response
-        self.metrics = metrics
     }
 
-    func map<U>(_ closure: (T) -> U) -> Response<U> {
-        Response<U>(value: closure(value), data: data, request: request, response: response, metrics: metrics)
+    func encode(to encoder: Encoder) throws {
+        try value.encode(to: encoder)
     }
 }
-
-#if swift(>=5.6)
-extension Request: @unchecked Sendable {}
-extension Response: @unchecked Sendable where T: Sendable {}
-#endif
