@@ -3,22 +3,29 @@
 // Copyright (c) 2021-2022 Alexander Grebenyuk (github.com/kean).
 
 import XCTest
-import Mocker
 #if canImport(FoundationNetworking)
 import FoundationNetworking
 #endif
 @testable import Get
 
 final class APIClientAuthorizationTests: XCTestCase {
-    
+    var client: APIClient!
+    private let delegate = MockAuthorizingDelegate()
+
+    override func setUp() {
+        super.setUp()
+
+        client = APIClient.github {
+            $0.delegate = delegate
+        }
+    }
+
     func testAuthorizationHeaderWidhValidToken() async throws {
         // GIVEN
-        let (client, delegate) = makeSUT()
-
         delegate.token = Token(value: "valid-token", expiresDate: Date(timeIntervalSinceNow: 1000))
         let url = URL(string: "https://api.github.com/user")!
         var mock = Mock.get(url: url, json: "user")
-        mock.onRequest = { request, arguments in
+        mock.onRequest = { request, _ in
             XCTAssertEqual(request.allHTTPHeaderFields?["Authorization"], "token valid-token")
         }
         mock.register()
@@ -26,37 +33,33 @@ final class APIClientAuthorizationTests: XCTestCase {
         // WHEN
         try await client.send(.get("/user"))
     }
-    
+
     func testAuthorizationHeaderWithExpiredToken() async throws {
         // GIVEN
-        let (client, delegate) = makeSUT()
-
         delegate.token = Token(value: "expired-token", expiresDate: Date(timeIntervalSinceNow: -1000))
         let url = URL(string: "https://api.github.com/user")!
         var mock = Mock.get(url: url, json: "user")
-        mock.onRequest = { request, arguments in
+        mock.onRequest = { request, _ in
             XCTAssertEqual(request.allHTTPHeaderFields?["Authorization"], "token valid-token")
         }
         mock.register()
-        
+
         // WHEN
         try await client.send(.get("/user"))
     }
 
     func testAuthorizationHeaderWithInvalidToken() async throws {
         // GIVEN
-        let (client, delegate) = makeSUT()
-
         delegate.token = Token(value: "invalid-token", expiresDate: Date(timeIntervalSinceNow: 1000))
         let url = URL(string: "https://api.github.com/user")!
         var mock = Mock(url: url, dataType: .json, statusCode: 401, data: [
             .get: "Unauthorized".data(using: .utf8)!
         ])
-        mock.onRequest = { request, arguments in
+        mock.onRequest = { request, _ in
             XCTAssertEqual(request.allHTTPHeaderFields?["Authorization"], "token invalid-token")
 
             var mock = Mock.get(url: url, json: "user")
-            mock.onRequest = { request, arguments in
+            mock.onRequest = { request, _ in
                 XCTAssertEqual(request.allHTTPHeaderFields?["Authorization"], "token valid-token")
             }
             mock.register()
@@ -65,22 +68,6 @@ final class APIClientAuthorizationTests: XCTestCase {
 
         // WHEN
         try await client.send(.get("/user"))
-    }
-
-    // MARK: - Helpers
-
-    private func makeSUT(using baseURL: URL? = URL(string: "https://api.github.com"),
-                         file: StaticString = #filePath,
-                         line: UInt = #line) -> (APIClient, MockAuthorizingDelegate) {
-        let delegate = MockAuthorizingDelegate()
-        let client = APIClient(baseURL: baseURL) {
-            $0.delegate = delegate
-            $0.sessionConfiguration.protocolClasses = [MockingURLProtocol.self]
-        }
-
-        trackForMemoryLeak(client, file: file, line: line)
-        
-        return (client, delegate)
     }
 }
 
@@ -98,8 +85,8 @@ private final class MockAuthorizingDelegate: APIClientDelegate {
 
         request.addValue("token \(token.value)", forHTTPHeaderField: "Authorization")
     }
-    
-    func shouldClientRetry(_ client: APIClient, for request: URLRequest, withError error: Error) async throws -> Bool {
+
+    func client(_ client: APIClient, shouldRetryRequest request: URLRequest, attempts: Int, error: Error) async throws -> Bool {
         if case .unacceptableStatusCode(let statusCode) = (error as? APIError), statusCode == 401 {
             token = try await tokenRefresher.refreshToken()
             return true
