@@ -3,26 +3,66 @@
 // Copyright (c) 2021-2022 Alexander Grebenyuk (github.com/kean).
 
 import XCTest
-import Mocker
 @testable import Get
 
 #if !os(Linux)
 final class APIClientSessionDelegateTests: XCTestCase {
+    private let delegate = SessionDelegate()
+    private var client: APIClient!
+
+    override func setUp() {
+        super.setUp()
+
+        self.client = .github {
+            $0.sessionDelegate = delegate
+        }
+    }
+
+    // MARK: - Configuration
+
+    func testThatDelegateQueueIsUsed() async throws {
+        // GIVEN
+        let delegateQueue = OperationQueue()
+        let queue = DispatchQueue(label: "com.get.api-client-tests")
+        let queueKey = DispatchSpecificKey<Void>()
+        queue.setSpecific(key: queueKey, value: ())
+        delegateQueue.underlyingQueue = queue
+
+        self.client = .github {
+            $0.sessionDelegate = delegate
+            $0.delegateQueue = delegateQueue
+        }
+
+        self.delegate.onMetrics = { [unowned self] _ in
+            XCTAssertNotNil(DispatchQueue.getSpecific(key: queueKey))
+            self.delegate.onMetrics = nil
+        }
+
+        // GIVEN
+        let url = URL(string: "https://api.github.com/user")!
+        Mock.get(url: url, json: "user").register()
+
+        // WHEN
+        try await client.send(.get("/user"))
+
+        // THEN
+        XCTAssertNil(self.delegate.onMetrics)
+    }
+
+    // MARK: - Global Delegate
 
     func testThatMetricsAreCollected() async throws {
         #if os(watchOS)
         throw XCTSkip("Mocker URLProtocol isn't being called for requests on watchOS")
         #endif
-        
-        // GIVEN
-        let (client, delegate) = makeSUT()
 
+        // GIVEN
         let url = URL(string: "https://api.github.com/user")!
         Mock.get(url: url, json: "user").register()
-        
+
         // WHEN
         try await client.send(.get("/user"))
-                                               
+
         // THEN
         XCTAssertEqual(delegate.metrics.count, 1)
         let metrics = try XCTUnwrap(delegate.metrics.first?.value)
@@ -30,28 +70,101 @@ final class APIClientSessionDelegateTests: XCTestCase {
         XCTAssertEqual(transaction.request.url, URL(string: "https://api.github.com/user")!)
     }
 
-    // MARK: - Helpers
+    // MARK: - Per-Task Delegate
 
-    private func makeSUT(using baseURL: URL? = URL(string: "https://api.github.com"),
-                         file: StaticString = #filePath,
-                         line: UInt = #line) -> (APIClient, SessionDelegate) {
-        let delegate = SessionDelegate()
-        let client = APIClient(baseURL: URL(string: "https://api.github.com")) {
-            $0.sessionConfiguration.protocolClasses = [MockingURLProtocol.self]
-            $0.sessionConfiguration.urlCache = nil
-            $0.sessionDelegate = delegate
+    func testSettingDelegate() async throws {
+        // GIVEN
+        self.client = .github()
+
+        final class MockDelegate: NSObject, URLSessionDataDelegate {
+            var response: URLResponse?
+
+            func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive response: URLResponse) async -> URLSession.ResponseDisposition {
+                self.response = response
+                return .cancel
+            }
         }
 
-        trackForMemoryLeak(client, file: file, line: line)
+        let url = URL(string: "https://api.github.com/user")!
+        Mock.get(url: url, json: "user").register()
 
-        return (client, delegate)
+        // WHEN
+        let delegate = MockDelegate()
+        let request = Request<Void>.get("/user")
+        do {
+            try await client.send(request, delegate: delegate)
+            XCTFail("Request was supposed to be cancelled")
+        } catch {
+            // Do nothing
+        }
+
+        // THEN
+        XCTAssertEqual(delegate.response?.url, url)
+    }
+
+    func testSettingDelegateCallbackBased() async throws {
+        // GIVEN
+        self.client = .github()
+
+        final class MockDelegate: NSObject, URLSessionDataDelegate {
+            var response: URLResponse?
+
+            func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive response: URLResponse, completionHandler: @escaping (URLSession.ResponseDisposition) -> Void) {
+                self.response = response
+                completionHandler(.cancel)
+            }
+        }
+        let url = URL(string: "https://api.github.com/user")!
+        Mock.get(url: url, json: "user").register()
+
+        // WHEN
+        let delegate = MockDelegate()
+        let request = Request<Void>.get("/user")
+        do {
+            try await client.send(request, delegate: delegate)
+            XCTFail("Request was supposed to be cancelled")
+        } catch {
+            // Do nothing
+        }
+
+        // THEN
+        XCTAssertEqual(delegate.response?.url, url)
+    }
+
+    func testSetTaskDelegateTogetherWithSessionDelegate() async throws {
+        // GIVEN
+        final class MockDelegate: NSObject, URLSessionDataDelegate {
+            var response: URLResponse?
+
+            func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive response: URLResponse, completionHandler: @escaping (URLSession.ResponseDisposition) -> Void) {
+                self.response = response
+                completionHandler(.cancel)
+            }
+        }
+        let url = URL(string: "https://api.github.com/user")!
+        Mock.get(url: url, json: "user").register()
+
+        // WHEN
+        let delegate = MockDelegate()
+        let request = Request<Void>.get("/user")
+        do {
+            try await client.send(request, delegate: delegate)
+            XCTFail("Request was supposed to be cancelled")
+        } catch {
+            // Do nothing
+        }
+
+        // THEN
+        XCTAssertEqual(delegate.response?.url, url)
     }
 }
 
 private final class SessionDelegate: NSObject, URLSessionTaskDelegate {
     var metrics: [URLSessionTask: URLSessionTaskMetrics] = [:]
-    
+    var onMetrics: ((URLSessionTaskMetrics) -> Void)?
+
     func urlSession(_ session: URLSession, task: URLSessionTask, didFinishCollecting metrics: URLSessionTaskMetrics) {
+        self.onMetrics?(metrics)
         self.metrics[task] = metrics
     }
 }
