@@ -16,10 +16,10 @@ public enum Paths {}
 
 extension Paths {
     public static var user: UserResource { UserResource() }
-    
+
     public struct UserResource {
         public let path: String = "/user"
-        
+
         public var get: Request<User> { .get(path) }
     }
 }
@@ -28,16 +28,16 @@ extension Paths {
 
 extension Paths.UserResource {
     public var emails: EmailsResource { EmailsResource() }
-    
+
     public struct EmailsResource {
         public let path: String = "/user/emails"
-        
+
         public var get: Request<[UserEmail]> { .get(path) }
-        
+
         public func post(_ emails: [String]) -> Request<Void> {
             .post(path, body: emails)
         }
-                
+
         public func delete() -> Request<Void> {
             .delete(path)
         }
@@ -50,10 +50,10 @@ extension Paths {
     public static func users(_ name: String) -> UsersResource {
         UsersResource(path: "/users/\(name)")
     }
-    
+
     public struct UsersResource {
         public let path: String
-                
+
         public var get: Request<User> { .get(path) }
     }
 }
@@ -62,10 +62,10 @@ extension Paths {
 
 extension Paths.UsersResource {
     public var followers: FollowersResource { FollowersResource(path: path + "/followers") }
-    
+
     public struct FollowersResource {
         public let path: String
-                
+
         public var get: Request<[User]> { .get(path) }
     }
 }
@@ -98,24 +98,23 @@ private final class GitHubAPIClientDelegate: APIClientDelegate {
     func client(_ client: APIClient, willSendRequest request: inout URLRequest) async throws {
         request.setValue("Bearer \("your-access-token")", forHTTPHeaderField: "Authorization")
     }
-        
+
     func shouldClientRetry(_ client: APIClient, for request: URLRequest, withError error: Error) async throws -> Bool {
         if case .unacceptableStatusCode(let status) = (error as? GitHubError), status == 401 {
             return await refreshAccessToken()
         }
         return false
     }
-    
+
     private func refreshAccessToken() async -> Bool {
         // TODO: Refresh (make sure you only refresh once if multiple requests fail)
         return false
     }
-    
+
     func client(_ client: APIClient, didReceiveInvalidResponse response: HTTPURLResponse, data: Data) -> Error {
         GitHubError.unacceptableStatusCode(response.statusCode)
     }
 }
-
 
 // MARK: - Usage
 
@@ -123,13 +122,13 @@ func usage() async throws {
     let client = APIClient(baseURL: URL(string: "https://api.github.com")) {
         $0.delegate = GitHubAPIClientDelegate()
     }
-    
-    let _ = try await client.send(Paths.user.get)
-    let _ = try await client.send(Paths.user.emails.get)
-    
+
+    _ = try await client.send(Paths.user.get)
+    _ = try await client.send(Paths.user.emails.get)
+
 //    try await client.send(Resources.user.emails.delete(["octocat@gmail.com"]))
-        
-    let _ = try await client.send(Paths.users("kean").followers.get)
-    
+
+    _ = try await client.send(Paths.users("kean").followers.get)
+
     let _: User = try await client.send(.get("/user")).value
 }
